requirements.txt
flask
sqlalchemy
pandas
numpy
nba_api
requests
python-dotenv